Updated dot spawning.
Dots now spawn according to the TileID of tiles on the map.  Update now checks the entire list of dots every frame for collision - hopefully there isn't a memory hole here or anything, cause if there is, it'll get ugly quickly.  (I am not good at javascript "foreach" functions, so if someone looks at the cludge of code there and thinks they can do it better, be my guest.)