Teleportation re-worked.
Teleportation now occurs via a "trigger" set for collision with tileID TELEPORT (i.e.:  3).  It changes the player's x and y values.  As of now, the only sanity checking it does is to make sure the player does not get teleported to a tile along the edge of the board - if we want to implement teleport tiles in the middle of the board, we'll need to add some more checks there to make that work because if the player lands *in* a teleport square, they teleport again the next time the update() function is called.