Implemented map transitions on "win."
I've implemented this by "destroying" the whole playing field when the last dot has been consumed and re-drawing the whole board, so this may introduce a whole bunch of bugs I can't foresee.  The good news is that if the bugs it introduces can be fixed without a huge amount of work, this ought to be scalable to as many maps as we want.